Renaming other error exports
